Split RoomProxyManager into separate files
